Created a simple client-server implementation of Roshambo
This initial version of Roshambo is a simple client-server command-line application built to explore how Socket.IO works and understand the rules and nuances of Roshambo.
Although named after the classic Rock-Paper-Scissors game, it extends to include the 'lizard' and 'Spock' moves, a variation popularized by the sitcom The Big Bang Theory.
The goal is to develop it into a full-fledged, production-grade online application. Stay tuned as this project unfolds..."